+++ Source/Game/Entities/Object/Update/UpdateFields.cs
@@ -3315,20 +3315,20 @@ namespace Game.Entities
     public class SpellPctModByLabel
     {
         public int ModIndex;
-        public float ModifierValue;
+        public double ModifierValue;
         public int LabelID;
 
         public void WriteCreate(WorldPacket data, Player owner, Player receiver)
         {
             data.WriteInt32(ModIndex);
-            data.WriteFloat(ModifierValue);
+            data.WriteFloat((float)ModifierValue);
             data.WriteInt32(LabelID);
         }
 
         public void WriteUpdate(WorldPacket data, bool ignoreChangesMask, Player owner, Player receiver)
         {
             data.WriteInt32(ModIndex);
-            data.WriteFloat(ModifierValue);
+            data.WriteFloat((float)ModifierValue);
             data.WriteInt32(LabelID);
         }
     }
@@ -3336,7 +3336,7 @@ namespace Game.Entities
     public class SpellFlatModByLabel
     {
         public int ModIndex;
-        public float ModifierValue;
+        public double ModifierValue;
         public int LabelID;
 
         public void WriteCreate(WorldPacket data, Player owner, Player receiver)
