+++ Source/Game/Scripting/Interfaces/IPlayer/IPlayerOnTakeDamage.cs
@@ -9,6 +9,6 @@ namespace Game.Scripting.Interfaces.IPlayer
     // Called when a player takes damage
     public interface IPlayerOnTakeDamage : IScriptObject, IClassRescriction
     {
-        void OnPlayerTakeDamage(Player player, float amount, SpellSchoolMask schoolMask);
+        void OnPlayerTakeDamage(Player player, double amount, SpellSchoolMask schoolMask);
     }
 }
