+++ Source/Game/Scripting/Interfaces/IUnit/IUnitModifyMeleeDamage.cs
@@ -7,6 +7,6 @@ namespace Game.Scripting.Interfaces.IUnit
 {
     public interface IUnitModifyMeleeDamage : IScriptObject
     {
-        void ModifyMeleeDamage(Unit target, Unit attacker, ref float damage);
+        void ModifyMeleeDamage(Unit target, Unit attacker, ref double damage);
     }
 }
